+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDRenderPipelineRuntimeResources.cs
@@ -37,6 +37,8 @@ namespace UnityEngine.Rendering.HighDefinition
             public Shader debugViewVirtualTexturingBlit;
             [Reload("Runtime/Debug/MaterialError.Shader")]
             public Shader materialError;
+            [Reload("Runtime/Debug/ProbeVolumeDebug.shader")]
+            public Shader probeVolumeDebugShader;
 
             // Lighting
             [Reload("Runtime/Lighting/Deferred.Shader")]
@@ -305,6 +307,8 @@ namespace UnityEngine.Rendering.HighDefinition
             public Shader SMAAPS;
             [Reload("Runtime/PostProcessing/Shaders/TemporalAntialiasing.shader")]
             public Shader temporalAntialiasingPS;
+            [Reload("Runtime/PostProcessing/Shaders/UpsampleScene.compute")]
+            public ComputeShader upsampleSceneCS;
             [Reload("Runtime/PostProcessing/Shaders/LensFlareDataDriven.shader")]
             public Shader lensFlareDataDrivenPS;
 
@@ -449,6 +453,8 @@ namespace UnityEngine.Rendering.HighDefinition
             public Mesh emissiveCylinderMesh;
             [Reload("Runtime/RenderPipelineResources/Mesh/Quad.fbx")]
             public Mesh emissiveQuadMesh;
+            [Reload("Runtime/RenderPipelineResources/Mesh/Sphere.fbx")]
+            public Mesh sphereMesh;
         }
 
         public ShaderResources shaders;