+++ com.unity.render-pipelines.universal/Runtime/2D/Renderer2DData.cs
@@ -1,7 +1,5 @@
 using System;
 using System.Collections.Generic;
-using UnityEngine.Rendering;
-using UnityEngine.Rendering.Universal;
 using UnityEngine.Scripting.APIUpdating;
 using UnityEngine.Serialization;
 
@@ -10,14 +8,14 @@ using UnityEditor;
 using UnityEditor.ProjectWindowCallback;
 #endif
 
-namespace UnityEngine.Experimental.Rendering.Universal
+namespace UnityEngine.Rendering.Universal
 {
     [Serializable, ReloadGroup, ExcludeFromPreset]
-    [MovedFrom("UnityEngine.Experimental.Rendering.LWRP")]
+    [MovedFrom("UnityEngine.Experimental.Rendering.Universal")]
     [HelpURL("https://docs.unity3d.com/Packages/com.unity.render-pipelines.universal@latest/index.html?subfolder=/manual/2DRendererData_overview.html")]
     public partial class Renderer2DData : ScriptableRendererData
     {
-        public enum Renderer2DDefaultMaterialType
+        internal enum Renderer2DDefaultMaterialType
         {
             Lit,
             Unlit,
@@ -75,11 +73,14 @@ namespace UnityEngine.Experimental.Rendering.Universal
         [SerializeField, Reload("Shaders/Utils/Sampling.shader")]
         Shader m_SamplingShader = null;
 
-        [SerializeField, Reload("Shaders/2D/ShadowGroup2D.shader")]
-        Shader m_ShadowGroupShader = null;
+        [SerializeField, Reload("Shaders/2D/Shadow2D-Projected.shader")]
+        Shader m_ProjectedShadowShader = null;
 
-        [SerializeField, Reload("Shaders/2D/Shadow2DRemoveSelf.shader")]
-        Shader m_RemoveSelfShadowShader = null;
+        [SerializeField, Reload("Shaders/2D/Shadow2D-Shadow-Sprite.shader")]
+        Shader m_SpriteShadowShader = null;
+
+        [SerializeField, Reload("Shaders/2D/Shadow2D-Unshadow-Sprite.shader")]
+        Shader m_SpriteUnshadowShader = null;
 
         [SerializeField, Reload("Shaders/Utils/FallbackError.shader")]
         Shader m_FallbackErrorShader;
@@ -91,8 +92,15 @@ namespace UnityEngine.Experimental.Rendering.Universal
         [HideInInspector]
         private Texture2D m_FallOffLookup = null;
 
+        /// <summary>
+        /// HDR Emulation Scale allows platforms to use HDR lighting by compressing the number of expressible colors in exchange for extra intensity range.
+        /// Scale describes this extra intensity range. Increasing this value too high may cause undesirable banding to occur.
+        /// </summary>
         public float hdrEmulationScale => m_HDREmulationScale;
         internal float lightRenderTextureScale => m_LightRenderTextureScale;
+        /// <summary>
+        /// Returns a list Light2DBlendStyle
+        /// </summary>
         public Light2DBlendStyle[] lightBlendStyles => m_LightBlendStyles;
         internal bool useDepthStencilBuffer => m_UseDepthStencilBuffer;
         internal Texture2D fallOffLookup => m_FallOffLookup;
@@ -102,9 +110,10 @@ namespace UnityEngine.Experimental.Rendering.Universal
         internal Shader pointLightVolumeShader => m_PointLightVolumeShader;
         internal Shader blitShader => m_BlitShader;
         internal Shader samplingShader => m_SamplingShader;
-        internal Shader shadowGroupShader => m_ShadowGroupShader;
-        internal Shader removeSelfShadowShader => m_RemoveSelfShadowShader;
         internal PostProcessData postProcessData { get => m_PostProcessData; set { m_PostProcessData = value; } }
+        internal Shader spriteShadowShader => m_SpriteShadowShader;
+        internal Shader spriteUnshadowShader => m_SpriteUnshadowShader;
+        internal Shader projectedShadowShader => m_ProjectedShadowShader;
         internal TransparencySortMode transparencySortMode => m_TransparencySortMode;
         internal Vector3 transparencySortAxis => m_TransparencySortAxis;
         internal uint lightRenderTextureMemoryBudget => m_MaxLightRenderTextureCount;
@@ -136,17 +145,18 @@ namespace UnityEngine.Experimental.Rendering.Universal
             normalsRenderTarget.Init("_NormalMap");
             shadowsRenderTarget.Init("_ShadowTex");
 
-            const int totalMaterials = 256;
-            if (shadowMaterials == null || shadowMaterials.Length == 0)
-                shadowMaterials = new Material[totalMaterials];
-            if (removeSelfShadowMaterials == null || removeSelfShadowMaterials.Length == 0)
-                removeSelfShadowMaterials = new Material[totalMaterials];
+            spriteSelfShadowMaterial = null;
+            spriteUnshadowMaterial = null;
+            projectedShadowMaterial = null;
+            stencilOnlyShadowMaterial = null;
         }
 
         // transient data
         internal Dictionary<uint, Material> lightMaterials { get; } = new Dictionary<uint, Material>();
-        internal Material[] shadowMaterials { get; private set; }
-        internal Material[] removeSelfShadowMaterials { get; private set; }
+        internal Material[] spriteSelfShadowMaterial { get; set; }
+        internal Material[] spriteUnshadowMaterial { get; set; }
+        internal Material[] projectedShadowMaterial { get; set; }
+        internal Material[] stencilOnlyShadowMaterial { get; set; }
 
         internal bool isNormalsRenderTargetValid { get; set; }
         internal float normalsRenderTargetScale { get; set; }