+++ com.unity.visualeffectgraph/Editor/GraphView/Views/VFXView.cs
@@ -445,6 +445,11 @@ namespace UnityEditor.VFX.UI
             compileButton.text = "Compile";
             m_Toolbar.Add(compileButton);
 
+            var resyncMatButton = new ToolbarButton(OnResyncMaterial);
+            resyncMatButton.style.unityTextAlign = TextAnchor.MiddleLeft;
+            resyncMatButton.text = "Resync Material";
+            m_Toolbar.Add(resyncMatButton);
+
             m_SaveButton = new ToolbarButton(OnSave);
             m_SaveButton.style.unityTextAlign = TextAnchor.MiddleLeft;
             m_SaveButton.text = "Save";
@@ -1409,6 +1414,11 @@ namespace UnityEditor.VFX.UI
             VFXViewWindow.currentWindow.autoCompile = !VFXViewWindow.currentWindow.autoCompile;
         }
 
+        void OnResyncMaterial()
+        {
+            controller.graph.Invalidate(VFXModel.InvalidationCause.kMaterialChanged);
+        }
+
         void OnCompile()
         {
             if (controller.model.isSubgraph)
@@ -1434,7 +1444,6 @@ namespace UnityEditor.VFX.UI
             foreach (var graph in graphToSave)
             {
                 graph.GetResource().WriteAsset();
-                graph.OnSaved();
             }
         }